Only kill the ssh agent if we started it from the script.

+++ remote/create_tunnel.sh
@@ -8,7 +8,7 @@ function cleanup() {
   if [ -d "${tmp_dir}" ]; then
     rm -rf "${tmp_dir}"
   fi
-  if [ ! -z "${SSH_AGENT_PID}" ]; then
+  if [ "${ssh_agent_launched}" = true ] && [ ! -z "${SSH_AGENT_PID}" ]; then
     kill ${SSH_AGENT_PID}
   fi
 }
@@ -17,6 +17,7 @@ SSHAGENT=$(which ssh-agent 2>/dev/null)
 SSHAGENTARGS="-s"
 if [ -z "${SSH_AUTH_SOCK}" ] && [ -x "${SSHAGENT}" ]; then
   eval $(${SSHAGENT} ${SSHAGENTARGS})
+  ssh_agent_launched=true
 fi
 
 user="${1}"
